fix parquet sorting error - #814
