+++ src/main/scala/org/broadinstitute/hail/utils/package.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.hail
 
 import java.net.URI
 
+import org.apache.hadoop.fs.PathIOException
 import org.apache.spark.AccumulableParam
 import org.broadinstitute.hail.check.Gen
 
@@ -183,6 +184,14 @@ package object utils extends Logging
   def genDNAString: Gen[String] = Gen.stringOf(genBase)
     .resize(12)
     .filter(s => !s.isEmpty)
+
+  def getPartNumber(fname: String): Int = {
+    val partRegex = ".*/?part-r-(\\d+)-.*\\.parquet.*".r
+    fname match {
+      case partRegex(i) => i.toInt
+      case _ => throw new PathIOException(s"invalid parquet file `$fname'")
+    }
+  }
 
   implicit def accumulableMapInt[K]: AccumulableParam[mutable.Map[K, Int], K] = new AccumulableParam[mutable.Map[K, Int], K] {
     def addAccumulator(r: mutable.Map[K, Int], t: K): mutable.Map[K, Int] = {
