+++ src/main/scala/org/broadinstitute/hail/utils/richUtils/RichSQLContext.scala
@@ -10,13 +10,6 @@ import org.broadinstitute.hail.utils._
 
 class RichSQLContext(val sqlContext: SQLContext) extends AnyVal {
   def readParquetSorted(dirname: String, selection: Option[Array[String]] = None): RDD[Row] = {
-    val partRegex = ".*/?part-r-(\\d+)-.*\\.parquet.*".r
-    def getPartNumber(fname: String): Int = {
-      fname match {
-        case partRegex(i) => i.toInt
-        case _ => throw new PathIOException(s"invalid parquet file `$fname'")
-      }
-    }
 
     val parquetFiles = sqlContext.sparkContext.hadoopConfiguration.globAll(Array(dirname + "/*.parquet"))
     if (parquetFiles.isEmpty)
