+++ src/test/resources/part-r-100001-8fd1217c-f5bf-4591-a4c5-08b1a812df09.gz.parquet
@@ -1,0 +1,1 @@
+second
